+++ UI/templates/home.html
@@ -7,11 +7,11 @@
     <!-- Bootstrap CSS -->
 
     <link rel="icon" type="image" href="static\social_icons\favicon.png" />
-    <script src="static\dependencies\jquery\3.6.0\jquery.min.js"></script>
-    <script src="static\dependencies\bootstrap-5.1.3-dist\js\bootstrap.min.js"></script>
+    <script src="static\dependencies\jquery.min.js"></script>
+    <script src="static\dependencies\bootstrap.min.js"></script>
     <link
       rel="stylesheet"
-      href="static\dependencies\bootstrap-5.1.3-dist\css\bootstrap.min.css"
+      href="static\dependencies\bootstrap.min.css"
     />
 
     <title>localGPT</title>
